fix: cap router packet allocation to read size
When the router received a packet from a WAN device and converted it
into a vec, the upperbound was not specified causing the whole 1610-byte
buffer to be allocated.  This manifested as a problem on the vhost-user
driver, which complained it could not write a full buffer to the device
(which had an MTU of 1500).  Fixed by specifying the upper bound on the
vector allocation.