refactor: PacketBuffers, new IPv4 types
To reduce allocations during program execution, a new PacketBufferPool
(and associated PacketBuffer struct) have been introduced. The intent is
to pre-allocate space on program start and contiuously use the same
(available) buffers, unless none are available, in which case a new
buffer is allocated.

Additionally, the IPv4 types have been redesigned.  There are now three
variants: Ipv4PacketOwned, Ipv4PacketRef, and Ipv4PacketMut.  In order,
the first takes ownership of the underlying buffer (aka a vec), the
second stores an immutable reference to a buffer (essentially providing
an Ipv4 view of a slice), and the last provides a mutable view into a
borrowed buffer.  Two traits have also been introducted, Ipv4Packet and
MutableIpv4Packet, that contain the expected methods for interacting
with Owned, Ref, or Mut IPv4 structures.